Units conversion for length inputs
Added units conversion, as per issue #21.  Seems to be working -
successfully converting from meters (in dynamo) to feet for the revit
api.  I tried changing from meters to feet in dynamo and nothing
happened - maybe you can't do that in the middle of a file - it has to
happen at the start?  More testing needed, but I think it's OK for now.